Do not report bogus/empty SMP cache_dir indexing stats
Non-disker kids in SMP instances with cache_dirs reported all-zero
indexing stats followed by a bogus claim that indexing took the whole
Unix epoch:

    2021/09/24 19:47:38 kid6| Finished rebuilding storage from disk.
        0 Entries scanned
        ...
        0 Swapfile clashes avoided
    Took 1632512858.62 seconds (0.00 objects/sec).

Once store_dirs_rebuilding reaches 1, we use counts.started() to check
whether the kid was responsible for indexing any cache_dirs. We do not
report bogus/empty indexing stats if the kid was not responsible.

TODO: This change preserves a storeDigestNoteStoreReady() call at the
end of indexing. Cache Digests probably do not work "as is" in SMP mode,
but fixing that is difficult and deserves a dedicated project.