Use exceptions to protect from certificate misinterpretation
See the previous commit for motivation. This change diff supports the
theory that the new CertGadgets functions are currently not used in any
code that must avoid exceptions. If such uses were hiding from us, the
previous commit would most likely not compile. A careful audit is still
warranted.

If future code erroneously uses CertGadgets functions in a context that
requires but lacks OpenSSL/GnuTLS, then it is better to throw than to do
nothing, even if that future code is not fully exception-safe.