Remove the SQUID_CHECK_OPENSSL_HELLO_OVERWRITE_HACK related code
This hack, allowed squid to stare-and-splice or peek-and-bump if
some conditions was true. In practice this was possible if the
web client and squid used the same openSSL release.

This hack code uses very low OpenSSL code, it is not working
with recent OpenSSL releases and it is disabled by default for
about five years.