Treat CertGadgets functions as optional information getters
... when there is no OpenSSL or GnuTLS support.

When OpenSSL or GnuTLS are enabled, these functions behave as if they
extract _optional_ certificate information (i.e. information that may
not be present in the certificate). The functions do not return
Optional<> (yet?), but their API does not treat missing information as
an error. If that design is correct, we should use it for the "missing
TLS library" cases as well -- it is not an error that we could not
extract some information.

Also fixed a few functions that were logging misleading "found..." debug
messages when they found nothing.