Bug #5158 - AnyP::Uri::host() mishandles [escaped] IPv6 addresses.
When an address is expected to be a string in the form "Host:Port", the
Host MUST be wrapped in square brackets iff it is a numeric IPv6 address.
The ":Port" part is usually optional. i.e. "[2001:db8::1]:443" or
"[2001:db8::1]".

There are 2 bugs relating to the handling of numeric IPv6 addresses:

Bug 1:
AnyP::Uri::host(const char *) is supposed to accept the host part of a
URI, but just copied it into hostAddr_ instead of using the fromHost()
method to set hostAddr_.  Since the argument is the host part of a URI,
numeric IPv6 addresses are wrapped in square brackets, which would
never be stripped and hostIsNumeric was therefore not set.  We now use
the fromHost() method to process the host name correctly.

Bug 2:
Conversely, AnyP::Uri::hostOrIp() is supposed to return the host name,
suitable for use in a URI or Host header, which means that numeric IPv6
addresses need to be wrapped in square brackets.  However, that
wrapping was never done.  We now use the toHostStr() method to correctly
format the string.

This also fixes the comment which describes IP::Address::toStr() - the
comment incorrectly said that toStr() creates a string containing the IP
address and port.  In fact, the resulting string is only the IP address
with no port, nor any square brackets that would be required to produce
a host:port format string.  The method to produce a Host:Port string is
toHostStr().