Alternate origin server selection.
When the proxy is making a direct connection to the origin server, it
will now connect to the host contained in the alt-host note, or
X-Alt-Host ICAP response header, if such a note / header exists. This
allows an external ACL or ICAP server to redirect the request to a
different host.  This is useful where services have a specific
"virtual IP" that clients can connect to to receive a modified
service.  e.g. this can be used to force the "Safe Search" feature on
for certain search engines.

This also fixes an existing bug:

The requests within peeked/bumped connections should usually be pinned so
that they go to the origin server via the original connection. However,
URIs which have been rewritten by ICAP or a url_rewrite_program, may
contain a different host name or port. In that case, it no longer makes
sense to pin the connection, and a new connection must be established to
the correct origin server.

In Squid 3.5, this worked as expected, but in Squid 5 and 6 it tries to
use the pinned connection and terminates the request since that
connection is no longer valid for the request.

I believe that this was broken by
https://github.com/squid-cache/squid/commit/daf80700130b6f98256b75c511916d1a79b23597
(Fix "BUG: Lost previously bumped from-Squid connection" (#460)).

There is already a redirected flag which records whether the URI has been
rewritten, but I have introduced a new redirected_origin flag to record
specifically whether the origin (i.e. host:port) has been changed so that
rewrites which only affect the path do not invoke new logic.

If the origin is rewritten, unpinConnection() is used to unpin the
original connection from the request. When deciding whether we need TLS
to the origin server, the redirect_origin flag is consulted. This is
because TLS to the origin server is usually required for any peeked /
bumped connections, but once it has been rewritten we need to rely on the
URI's scheme instead, since the rewritten URI may be to an HTTP resource
instead of HTTPS.

It is not clear to me why redirected_origin needs to be checked in
HttpRequest::pinnedConnection(). unpinConnection() doesn't completely
undo pinConnection() - e.g. it intentionally doesn't clear the
pinning.pinned flag, and HttpRequest::pinnedConnection() seems to repin
the connection after it has been unpinned, so the redirect_origin check
there prevents that.