fixup: Fixed stale, misplaced, and misspelled new method descriptions
The "previous generation" messages are not the reason we need these
AtStart() methods! Those stale messages are handled by the existing
methods (that the new methods call).

The new methods are needed because the old code assumes that the queue
state reflects our process state: If our reader flag is set, then we do
not need notifications about new queued items. This assumption is false
after a reader restart.

XXX: I cannot think of a good way to avoid documentation duplication
here. We should not use \copydoc A in method B here because method A is
unrelated to B; those changing A do not need to know about B. And if we
add "See also: B" to method A, then \copydoc will copy that note to B!