Fix build on openbsd 7.0
OpenBSD doesn't offer CPU_SET and CPU_ISSET. Implement
their stubs as inline functions to give the compiler
proper hints about arguments (non) use

OpenBSD's STL has a const-correctness bug in
std::unordered_map when supplying an allocator.
Implement an autoconf test for this case and
do not use allocators in that case.

Tested working on openbsd in
https://build.squid-cache.org/job/anybranch-openbsd-matrix/11/
Tested non-regression on Linux in
https://build.squid-cache.org/job/anybranch-amd64-matrix/45/