fixup: GnuTLS modifies certificate struct when searching for components
We have to make our internal API worse to accommodate GnuTLS. I have
considered hiding this modification behind a const_cast instead, but I
think that can (and perhaps should) be done only if we are certain that
GnuTLS modifications are an equivalent of modifying a mutable data
member and, hence, can be safely ignored without introducing bugs. My
very quick search did not find any relevant explanations or assurances.

https://gitlab.com/gnutls/gnutls/blob/ba477f5/lib/x509/x509.c#L2798-2814

Let's see how far we get with a non-constant Certificate. It was not
constant before the conversion from pointer (i.e. the previous branch
commit) so hopefully it will not be such a big deal long-term. It does
make our APIs look odd, so I added an explanation. We may need to find
a better way to explain that (without duplicating the same phrase).