Preserve configured order of intermediate CA certificate chain
    https_port ... tls-cert=signing,itsIssuer,itsIssuerIssuer.pem

The order was reversed in commit cf48712, probably by accident. Wrong
order violates TLS protocol and breaks TLS clients that are incapable of
reordering received intermediate CAs. Squid deployments that use
wrong-order bundles (to compensate for this bug) should reorder their
bundles when deploying this fix (or wait for Squid to order certificates
correctly, regardless of the bundle order -- a work in progress).

OpenSSL sends the signing certificate first. After that, OpenSSL sends
certificates in the order they are stored in the chain, so we must push
them in on-the-wire order, as defined by RFC 8446 Section 4.4.2: "The
sender's certificate MUST come in the first CertificateEntry in the
list. Each following certificate SHOULD directly certify the one
immediately preceding it."