fixup: Use a more specific parameter name
This code is difficult to read because getXxRecord() does not have a
parameter to supply the sharedName. AFAICT, instead of asking for the
name, getXxRecord() uses a copy of the previously configured
masterx_shared_name, which usually(XXX) matches the masterxSharedValue()
parameter, which is the _current_ configured masterx_shared_name.

However, this PR is about the GCC warning, not other non-trivial fixes.