Inform admins using -X about (the effects of) their -d (non-)usage
    cache.log: Using -X without -d; stderr will see no messages
    stderr: Using -X and -d1; stderr will not see some messages

This information is primarily useful when admins expect -X to force -d9
and are surprised by the lack of stderr debugging. This information is
often the first message logged, increasing the chance that an admin will
notice it.

There is no `else` clause because

* an early level-2+ message will not be logged (without -X) anyway
* logging a level-0/1 message just to record -d setting is excessive

Also added a -d9 hint to -X documentation in `squid -h`.