Bug 5186: noteDestinationsEnd check failed: transportWait
When the "no more destinations to try" notification comes after the last
forwarding/tunneling attempt has failed, the !destinationsFound block
does not run (because destinations were found), the usingDestination()
block does not run (because we are done with that last/failed
destination), but transportWait is false (for the same reason).

XXX: Fix FwdState::noteDestinationsEnd() as well, re-unifying the code.

TODO: Can we assert that savedError is nil inside the !destinationsFound
block? The old "all found paths have failed" reason does not makes sense
there because there were no paths found if destinationsFound is false,
but is there another way to get inside that block with non-nil
savedError?