fixup: Polished primary TrafficMode method descriptions, placement
Grouped the three primary port mode reporting methods together.

Polished forwarded() description to explain why it is negating the
interceptedSomewhere() even though nobody can guarantee the lack of
interception and Squid cannot detect some forms of interception. Both
interceptedSomewhere() and the forwarded() methods are _not_ reporting
(or implying) whether the traffic was intercepted. They report (or
imply) whether Squid should behave as if the traffic was (or was not)
intercepted on the way to the port.

Simplified interceptedSomewhere() description.

Use (more of) draft-ietf-httpbis-semantics terminology.