fixup: Renamed TrafficMode::forwarded() to explicitProxy()
draft-ietf-httpbis-semantics calls these kind of proxy configurations
just "proxy", but we probably should not do that in Squid context.

We need to decide whether we should be naming this methods to describe
the port itself or traffic that the port expects. The port is an
"explicit proxy" port, but traffic is "[explicitly] forwarded" (or some
better but similarly structured/worded phrase). Since these methods are
usually called in "port.flags.method" context, it may be better to
describe the port rather than traffic. However, it may be difficult to
find a good descriptive phrase for a port that handles intercepted
(somewhere) traffic without necessarily being an intercepting proxy
port!