Removed unused (and problematic) RegexPattern move optimizations
... to simplify.

Technically, calling POSIX regfree() with a zeroed regex is a UB because
regfree() is only defined for compiled regexes. Squid ignores that
problem on other code paths as well, but since these move optimizations
are actually unused (and are not expected to be used), we can just
delete one problematic code path and simplify the (future) UB fix.