fixup: Synced RegexData cfg code with throwing RegexPattern ctor
This branch commit restores the (problematic) official code
functionality -- skipping invalid REs.

I removed "XXX: old code did this. It looks wrong" code/comments during
this refactoring. The comment (added in commit e2b7452) can be
considered correct: Out of the 3-4 situations where aggregated REs can
be empty, the "old" odd-looking code handled only cases 1 and 4.

1. "-i/+i" at the beginning of the ACL data: No accumulated REs yet.
2. "-i/+i" after a buffering/joining overflow emptied accumulated REs.
3. "regex1 +i -i regex2" and similar sequences: Accumulated REs (regex1)
   are emptied upon seeing "+i". Squid then tries to compile still-empty
   accumulated REs upon seeing the "-i".
4. An empty RE token. This may not be possible with recent RE parsers.

The branch code no longer calls compileREs() with an empty list of REs,
handling situations 1-3. I let the regex library handle the empty
(zero-length) RE token instead of treating it as a special case.

It is possible to refactor the code to avoid code duplication
surrounding compileREs() calls, but that duplication existed before this
branch changes and such refactoring would require a lot of changes
unrelated to this branch, so it is best done in a dedicated PR.