Do not report REs optimization failure when an individual RE is broken
When an individual RE is broken, compileOptimisedREs() failure is most
likely not an optimization failure, and even if it is, it is irrelevant:
We should not be optimizing broken REs.

Yes, we (try to) optimize all REs, valid and not, but we do that just
because we choose to ignore the fact that broken REs can be "optimized"
into a completely different[1] valid combined RE:

    compileOptimisedREs: adding RE 're)?((1'
    compileOptimisedREs: adding RE '2))+(re'
    RegexPattern: (re)?((1)|(2))+(re)
    compileOptimisedREs: 2 REs are optimised into one RE.

[^1]: It is impossible to meaningfully compare with broken REs, but we
can assume that the combined version was _not_ the REs author intent!