fixup: Polished coredump_dir side effects documentation
Squid actions are not exactly OS-specific but depend on certain APIs
availability instead.

The details of what happens are too varying and environment-specific to
document in squid.conf IMO, but I left references to APIs that can be
easily looked up on the web and used as starting points. If the admin
needs exact answers, they will need to research Squid code corresponding
to their environment (which these API references make easy to find in
Squid sources).

I also think that, given the current implementation, it is important to
tell the admin what happens if Squid does not know how to adjust the
process permissions. The admin will have no way of "fixing" that ERROR,
even if their Squid successfully generates/leaves core dumps.