fixup: Polished Assure() description
The "does not create a threat" phrasing was too strong: Any exception or
other non-trivial action is a potential threat. It would be virtually
impossible to prove non-existence, and we should not expect such proofs.

Also, "go unnoticed" is not the defining characteristics of the
"dangerous state" but just one (easy-to-understand) possibility.