fixup: Removed ENOSYS, reduced growing code duplication, detailed errors
Error reporting sample:

    2022/02/09 12:14:29| ERROR: Cannot make the process traceable:
        exception: prctl(PR_SET_DUMPABLE) failure: (1) Operation not...
        exception location: tools.cc(289) operator()

Instead of reporting an ENOSYS error when we do not know how to make the
process traceable in a given environment, let's assume that the process
is already traceable enough (by default) to dump cores in such
environments.

Alternatively, we could make that special message a level-1 warning and
assign a debugs() message ID to it so that admins that are bothered by
it would disable it via cache_log_message. It is not clear which
alternative is best.