fixup: Decided how to handle early errors inside noteDestinationsEnd()
These errors are probably impossible in the current code because any
early error would preclude us from starting selecting destinations.
However, there is no good reason to handle them specially in FwdState
because noteDestinationsEnd() does the same thing regardless of whether
the error is early or not.

Tunnel.cc code is messier because we must create an error before calling
sendError(). However, when we merge TunnelStateData and FwdState, we may
be able to follow the simpler FwdState design here.

This tiny change can be considered in scope because the now-adjusted
code had a false "all found paths have failed" implication. If that
statement were true, the primary bug fix could have been different.