Extend TunnelStateData::usingDestination() state while server closes
server.conn->close() makes Comm::IsConnOpen() false immediately, but the
tunnel is still running and may even receive new destination paths. This
change does not solve all the problems, but it is a step in the right
direction -- extending usingDestination() state to cover periods when we
are not ready to try the next destination.