Extend TunnelStateData::usingDestination() state while client writes
... the last from-server chunk, after the server connection closes.

With this commit, tunnel.cc is on par with FwdState as far as Bug 5090
protections are concerned. Tunnel.cc never retries successfully
established tunnels, which simplifies transporting() maintenance
compared to FwdState, but tunnels may start reforwarding in most cases,
and most of the code does not need to know about this (temporary?)
simplification.

Also stop receiving possible spare destination paths if we are not going
to use them (because retriable became false).

Also use chronological order of with-path noteDestination() conditions.
FwdState::noteDestination() already uses that order.