Detach libsbuf from StatHist to facilitate SBuf reuse
The stated purpose of sbuf/DetailedStats was to "avoid adding external
dependencies to the SBuf code". DetailedStats failed to accomplish that
because it introduced a dependency on StatHist (which depends on Store).

To break the unwanted dependency, we outsource at-destruction-time size
statistics collection to external-to-SBuf code, configurable via the new
SBufStats fields. TODO: Long-term, StatHist itself should be refactored
to become reusable without unwanted heavy dependencies.