Refactor Half-Closed FD monitoring logic
The complex event handler to determine when half-closed FDs have
lost their read handler can be replaced with a fd_table flag set
when the mode is enabeld and unset on disable.

This flag can be tested by the read handler to re-schedule
monitoring correctly even when the mdode is disabled while that
Call is pending in the Async dialer queue.

It also provides a far simpler method for other comm related
logic to test for write-only (half-closed) status without the
current complex registry lookup.

The DescriptorSet registry already duplicated the fd_table
halfClosedReader member. Now with the write_only flag it is
fully redundant and can be removed.