Simplified further, completing the XXX fix started in previous commit
Removed an "extra" data connection check that was leaving stale true
Ftp::Client::data.read_pending setting (at least). AFAICT, we now came
back much closer to the official code as far as handling of the various
flags across delayed reads is concerned. That handling is far from
perfect but we do not want to be responsible for polishing that complex
code (and there is already another project doing that polishing).