Keep ServerBump::serverSession (i.e. TLS Squid-peer connection) fresh
The session may change if, for example, Squid retries using a different
destination after failing to finish the first TLS negotiation attempt:

    2022/05/06 14:02:56.887| 83,5| Io.cc(92) Handshake: 0/0
    2022/05/06 14:02:56.887| 83,7| ErrorDetail: got 0x14094410
    error:14094410:...:ssl3_read_bytes:sslv3 alert handshake failure

I speculate that the erroneous `if` statement was an attempt to avoid
problems with self-assignment. Instead, we should rely on our
Security::SessionPointer handling self-assignment correctly, of course.