fixup: Propagate API changes to callers
... while avoiding exponential growth of changed functions by wrapping
code in calling chains that are not yet ready to handle exceptions (and
cannot be easily converted) with try-return-true/catch blocks.

I also changed the old ReadX509Certificate() name to make sure we cover
all the callers, including those we might get through merging with
then-current master. The new name is also simpler. Other names are not
yet consistent regarding using X509; we should migrate towards the
simpler spellings that do not using that "Certificate" augmentation.