fixup: Do not report successfully extracted certificate fields
... because known callers already report them anyway OR should report
them (and probably will do so eventually).

Alternatively, we could create pairs or parameterized extractors so that
the caller can decide whether they want the extractor to dump the
extracted information to cache.log as a side effect. However, that adds
work for us now and for the callers in the future so lets KISS for now.