fixup: Removed a likely misleading TODO
It would be nice to avoid this memory allocation, and OpenSSL may have
enough knobs for doing so (Christos wrote a sketch that may be
sufficient), but the amount of special code this optimization will
require will be significant. And once we add GnuTLS support (if the same
optimization is possible there at all), we would be looking at a rather
complex optimization with little performance gain (at least based on the
current Certificate printing use cases).

A much better strategy may be to parameterize both TLS libraries to
proxy their memory allocations through Squid's pooling-capable libmem!