Do not send two copies of the self-signed signing certificate
A single-certificate chain containing a self-signed certificate R was
sent as R-R (instead of just R).

The original "will not be chained" debugs() message contents for the
self-signed signing certificate felt odd because the signing
certificates should never be added to the chain, regardless of whether
they are self-signed. What we probably meant to say is that Squid will
not form/sent a chain of CA certificates after the signing certificate
because/when that signing certificate is self-signed.