Keep Ip::Intercept 'active' flags consistent with ./configure options
... and reject such inconsistent configurations.

Before this change we could, for example, run Squid with a port marked
as 'intercept' but without interception enabled by ./configure.  This
would end up having a non-operational port rejecting all transactions
with a level-1 message. Is seems that a better alternative is to forbid
such invalid configurations at all.