Collect all official CONTRIBUTORS we can before manual pruning
This dirty CONTRIBUTORS snapshot contains all contributors that I could
find and add from the following (master and v3+ branches) sources:

* All commit authors
* All commit co-authors (from Co-authored-by entries)
* All CONTRIBUTORS file revisions (because we lost some entries)

I did not add a few non-ASCII names, but their emails or even full
entries are present or, in one case where email was not available, I
added an ASCII rendering of their name (its UTF-8 version was garbled
earlier in this branch so this can be considered an improvement).

This commit adds a few previously lost entries. I hope it does not lose
any entries.

I am committing this intermediate result to show what has been changed
compared to the previous state of CONTRIBUTORS on this branch. The next
step would be to manually polish these entries so that this file can
be used by automated scripts, as a reference point going forward.