Manual polishing of CONTRIBUTORS entries
Deleted known duplicates and a couple of arguably useless entries.

Added or polished a couple of known names to/in the existing entries.

Needs more work, but we do not need to get all of it now, in this
branch. This effort can be incremental (and mistakes can be
undone/fixed) because we will never re-generate raw CONTRIBUTORS from
scratch once this branch is merged.

This commit is a reference point for automated CONTRIBUTORS updates.