fixup: The branch code was asserting when decoding tls-dh
The related OpenSSL API is poorly documented and RAII-unfriendly. The
fixed code is based on OpenSSL code analysis (including a few
OSSL_DECODER_CTX_new_for_pkey() callers) and Squid tests. The current
version is one of the simplest ones I can think of that would still
allow a straightforward conversion to exceptions later.

If I misjudged the OpenSSL intent, then this code may eventually assert,
and we will need to make it more complex (and robust).