fixup: Avoid rawCtx pointer by adjusting dctx declaration
The unusual (for the current Squid code) {braces} inside the
if-statement condition are required for the code to compile:

    ServerOptions.cc:393:34: error: expected initializer before ‘(’
    393 |     if (const DecoderContext dctx(...)) {
        |                                  ^

I suspect that, without them, (the beginning of) the condition is
interpreted as a function declaration.

This can be further simplified/improved by creating a C++-friendly
OSSL_DECODER_CTX_new_for_pkey() wrapper (that would return a smart ptr).