Found a way to allow the compiler to force SwapMetaTypeMax updates
We cannot ask the compiler to compute SwapMetaTypeMax value without
screwing up our switch statements, but we can ask the compiler to tell
us when we forgot to update it. See -Wswitch, enabled by -Wall.

Until C++17, this requires converting a static assertion[1] into a
dynamic one, but perhaps it is an acceptable temporary trade-off.

[1] Two duplicated assertions, to be precise.

Context:
https://github.com/squid-cache/squid/pull/1067/files#r889976450
https://cpp.godbolt.org/z/oYnhn644r