Avoid (ab)using STORE_META_VOID as a start of serialized ID ranges
STORE_META_VOID is a SwapMetaType ID. Conceptually, it does not have a
serialized value at all because it is never serialized and should never
be serialized. Comparing/setting RawSwapMetaType values with/to
STORE_META_VOID is essentially an "incompatible type" error (that the
compiler cannot detect for us, unfortunately).