fixup: Do not duplicate Extract() across swap meta reading code
The now-shared SwapMetaExtract() is in SwapMetaView.h because we do not
have a better header for this helper function. Moving it to SwapMetaIn.h
would be worse because it would expose high-level Unpack*() callers to
this low-level helper function. We could create a dedicated file for
low-level unpacking gadgets, but that sounds like an overkill for just
one function that has an imperfect place already. We also could merge
SwapMetaView back into SwapMetaIn.cc, but that is likely trigger a
change request to isolate that helper class.