fixup: More polishing
I removed the "Skip this StoreEntry" TODO because that decision should
probably not be done based on meta.type, and I am not even sure it is a
good idea to skip at least some of the entries with legacy fields. This
whole idea requires further study and I do not want this rather specific
branch TODO to (mis)direct that study.